requirements.txt
Flask==0.10.1
Flask-Assets==0.8
Flask-Compress==0.10.0
Flask-Mail==0.9.0
Flask-OpenID==1.1.1
Flask-Oauthlib==0.4.2
Flask-Redis==0.0.3
Flask-SQLAlchemy==1.0
Flask-Uploads==0.1.3
Flask-WTF==0.9.2
Flask-gzip==0.1
Jinja2==2.7.1
MarkupSafe==0.18
SQLAlchemy==0.8.2
WTForms==1.0.5
Werkzeug==0.9.4
beautifulsoup4==4.3.2
blinker==1.3
boto==2.13.3
#distribute==0.6.28
gunicorn==18.0
httplib2==0.8
itsdangerous==0.23
oauth2==1.5.211
psycopg2==2.5.1
pyOpenSSL==0.13.1
pysolr==3.1.0
python-openid==2.2.5
python-linkedin==4.0
redis==2.9.0
requests==2.0.0
requests-oauth==0.4.1
requests-oauthlib==0.4.0
stripe==1.9.5
webassets==0.8
wsgiref==0.1.2
wtf==0.1
